+++ src/queue/utils/constants.ts
@@ -1,7 +1,7 @@
 import { StoreDestinationArray } from "../../common/persistence/IExtentStore";
 
-export const VERSION = "3.22.0";
-export const QUEUE_API_VERSION = "2021-10-04";
+export const VERSION = "3.23.0";
+export const QUEUE_API_VERSION = "2022-11-02";
 export const DEFAULT_QUEUE_SERVER_HOST_NAME = "127.0.0.1"; // Change to 0.0.0.0 when needs external access
 export const DEFAULT_QUEUE_LISTENING_PORT = 10001;
 export const IS_PRODUCTION = process.env.NODE_ENV === "production";
@@ -90,6 +90,7 @@ export const DEFAULT_QUEUE_PERSISTENCE_ARRAY: StoreDestinationArray = [
 ];
 
 export const ValidAPIVersions = [
+  "2022-11-02",
   "2021-12-02",
   "2021-10-04",
   "2021-08-06",