+++ tests/table/apis/table.entity.query.test.ts
@@ -981,7 +981,7 @@ describe("table Entity APIs test - using Azure/data-tables", () => {
     await tableClient.deleteTable();
   });
 
-  it("should find a property identifier starting with underscore, @loki", async () => {
+  it("15. should find a property identifier starting with underscore, @loki", async () => {
     const tableClient = createAzureDataTablesClient(
       testLocalAzuriteInstance,
       getUniqueName("under")
@@ -1001,6 +1001,171 @@ describe("table Entity APIs test - using Azure/data-tables", () => {
       })
       .next();
     assert.notStrictEqual(queryResult.value, undefined);
+    await tableClient.deleteTable();
+  });
+
+  // issue 1828
+
+  it("16. should find guids when using filter with ge, lt, gt and ne, @loki", async () => {
+    const tableClient = createAzureDataTablesClient(
+      testLocalAzuriteInstance,
+      getUniqueName("guidfilters")
+    );
+    const partitionKeyForQueryTest = getUniqueName("1");
+    const totalItems = 3;
+    await tableClient.createTable();
+
+    const guidEntities: TableTestEntity[] = [];
+
+    const entity0 = entityFactory.createBasicEntityForTest(
+      partitionKeyForQueryTest
+    );
+    entity0.guidField.value = "11111111-1111-1111-1111-111111111111";
+    guidEntities.push(entity0);
+
+    const entity1 = entityFactory.createBasicEntityForTest(
+      partitionKeyForQueryTest
+    );
+    entity1.guidField.value = "22222222-2222-2222-2222-222222222222";
+    guidEntities.push(entity1);
+
+    const entity2 = entityFactory.createBasicEntityForTest(
+      partitionKeyForQueryTest
+    );
+    entity2.guidField.value = "33333333-3333-3333-3333-333333333333";
+    guidEntities.push(entity2);
+
+    for (let i = 0; i < totalItems; i++) {
+      const result = await tableClient.createEntity(guidEntities[i]);
+      assert.notStrictEqual(result.etag, undefined);
+    }
+
+    const maxPageSize = 10;
+    let testsCompleted = 0;
+    const queriesAndExpectedResult = [
+      {
+        index: 0,
+        queryOptions: {
+          filter: odata`(PartitionKey eq ${partitionKeyForQueryTest}) and (guidField ne guid'${guidEntities[1].guidField.value}')`
+        },
+        expectedResult: 2, // we insert 3 and don't want this one
+        expectedValue: guidEntities[0].guidField.value
+      },
+      {
+        index: 1,
+        queryOptions: {
+          filter: odata`(PartitionKey eq ${partitionKeyForQueryTest}) and (guidField lt guid'${guidEntities[1].guidField.value}')`
+        },
+        expectedResult: 1,
+        expectedValue: guidEntities[0].guidField.value
+      },
+      {
+        index: 2,
+        queryOptions: {
+          filter: odata`(PartitionKey eq ${partitionKeyForQueryTest}) and (guidField ge guid'${guidEntities[1].guidField.value}')`
+        },
+        expectedResult: 2,
+        expectedValue: guidEntities[1].guidField.value
+      },
+      {
+        index: 3,
+        queryOptions: {
+          filter: odata`(PartitionKey eq ${partitionKeyForQueryTest}) and (guidField gt guid'${guidEntities[1].guidField.value}')`
+        },
+        expectedResult: 1,
+        expectedValue: guidEntities[2].guidField.value
+      }
+    ];
+
+    for (const queryTest of queriesAndExpectedResult) {
+      const entities = tableClient.listEntities<TableTestEntity>({
+        queryOptions: queryTest.queryOptions
+      });
+      let all: TableTestEntity[] = [];
+      for await (const entity of entities.byPage({
+        maxPageSize
+      })) {
+        all = [...all, ...entity];
+      }
+      assert.strictEqual(
+        all.length,
+        queryTest.expectedResult,
+        `Failed with query ${queryTest.queryOptions.filter}`
+      );
+      if (all[0] !== undefined) {
+        all.sort((a, b) => {
+          return (
+            parseInt(a.guidField.value[1], 10) -
+            parseInt(b.guidField.value[1], 10)
+          );
+        });
+        assert.strictEqual(
+          all[0].guidField.value,
+          queryTest.expectedValue,
+          `Test ${queryTest.index}: Guid value ${all[0].guidField.value} was not equal to ${queryTest.expectedValue} with query ${queryTest.queryOptions.filter}`
+        );
+      } else {
+        assert.strictEqual(
+          all[0],
+          queryTest.expectedValue,
+          `Value ${all[0]} was not equal to ${queryTest.expectedValue} with query ${queryTest.queryOptions.filter}`
+        );
+      }
+
+      testsCompleted++;
+    }
+    assert.strictEqual(testsCompleted, queriesAndExpectedResult.length);
+    await tableClient.deleteTable();
+  });
+
+  it("17. should work correctly when query filter single boolean and partition filter, @loki", async () => {
+    const tableClient = createAzureDataTablesClient(
+      testLocalAzuriteInstance,
+      getUniqueName("partwithnot")
+    );
+
+    await tableClient.createTable();
+
+    const result1 = await tableClient.createEntity({
+      partitionKey: `Part1`,
+      rowKey: `1`,
+      foo: "testEntity1"
+    });
+    assert.notStrictEqual(result1.etag, undefined);
+
+    const result2 = await tableClient.createEntity({
+      partitionKey: `Part2`,
+      rowKey: `1`,
+      foo: "testEntity2"
+    });
+    assert.notStrictEqual(result2.etag, undefined);
+
+    const result3 = await tableClient.createEntity({
+      partitionKey: `Part3`,
+      rowKey: `1`,
+      foo: "testEntity3"
+    });
+    assert.notStrictEqual(result3.etag, undefined);
+
+    const maxPageSize = 5;
+    const entities = tableClient.listEntities<TableEntity<{ foo: string }>>({
+      queryOptions: {
+        filter: odata`not (PartitionKey lt 'Part2')`
+      }
+    });
+    let all: TableEntity<{ foo: string }>[] = [];
+    for await (const entity of entities.byPage({
+      maxPageSize
+    })) {
+      all = [...all, ...entity];
+    }
+    assert.strictEqual(all.length, 2);
+    all.forEach((entity) => {
+      assert.ok(
+        entity.partitionKey === `Part2` || entity.partitionKey === `Part3`
+      );
+    });
+
     await tableClient.deleteTable();
   });
 });