+++ ChangeLog.md
@@ -4,9 +4,32 @@
 
 ## Upcoming Release
 
+Blob:
+
+- Fixed issue of: blob batch subresponse is slightly different from the on from Azure serivce, which causes exception in CPP SDK.
+- Fixed issue of: setMetadata API allows invalid metadata name with hyphen.
+
+## 2023.03 Version 3.23.0
+
+General:
+
+- Return 404 StatusCode when Storage account not exist
+- Migrated tslint to eslint.
+- Typescript upgraded from 4.2.4 to 4.9.5.
+- Migrated test pipeline from Node.js 10/12 to Node.js 14/16/18.
+- Bump up service API version to 2022-11-02
+
+Blob:
+
+- Fixed issue for user delegation key when uploading a blob from a container SAS
+- Upgraded swagger spec to API version 2021-10-04.
+
 Table:
 
 - Fixed issue for querying on identifiers starting with underscore.
+- Corrected query parsing logic for single boolean terms.
+- Fixed issue for querying GUIDs using operators other than eq and ne
+- GUID queries only support persistent storage on legacy (string) format GUIDs for eq and ne operators, other operators will only evaluate newly stored entities.
 
 Queue: